backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sqlmodel==0.0.22          # اگر sqlmodel استعمال کر رہے ہو تو یہ رکھیں
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.9
httpx==0.27.2
pydantic[email]==2.9.2
email-validator==2.2.0
psycopg2-binary==2.9.10
uuid==1.30
groq==0.4.1          